Should be a good base of work.
TCOD is chewing a nice part of the work.

Now we need a scene system to package objects, a way to transit objects
and states between scenes, and a way to load scenes, that should include
an additive way. We should also be able to describe scenes in an
external file, something like XML or YML.

We need a way to configure the engine via an external file as well.